projects/genomicsdb-schema-manager/pyproject.toml
[project]
name = "genomicsdb-schema-manager"
authors = [
    { name = "fossilfriend", email = "egreenfest@gmail.com" },
]
dependencies = [
    "alembic (>=1.15.2,<2.0.0)",
    "alembic-utils (>=0.8.8,<0.9.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "nh3 (>=0.2.21,<0.3.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "strenum (>=0.4.15,<0.5.0)",
]
requires-python = ">=3.11,<4.0"
version = "0.1.0"

[tool.poetry]
packages = [
    { from = "../../components", include = "niagads/common" },
    { from = "../../components", include = "niagads/database" },
    { from = "../../components", include = "niagads/enums" },
    { from = "../../components", include = "niagads/exceptions" },
    { from = "../../components", include = "niagads/genomics" },
    { from = "../../components", include = "niagads/genomicsdb" },
    { from = "../../components", include = "niagads/settings" },
    { from = "../../components", include = "niagads/utils" },
    { from = ".", include = "helpers" },
]

[tool.tomlsort]
all = true
in_place = true
spaces_indent_inline_array = 4

[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
    "poetry-core>=1.0.0",
]
